# caliborn.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "caliborn",
    "description": "API server for lumiRadio",
    "contact": {
      "name": "cozygalvinism"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users/{id}": {
      "get": {
        "tags": [
          "crate::routes::admin::users"
        ],
        "operationId": "get_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "crate::routes::admin::users"
        ],
        "operationId": "patch_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
...
    "/user/me": {
      "get": {
        "tags": [
          "crate::routes::user"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "User was successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "An authorization error occurred (e.g. invalid access token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "message": "Invalid access token"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Internal Server Error",
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/user/me/pay": {
      "post": {
        "tags": [
          "crate::routes::user"
        ],
        "operationId": "pay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Boonbucks transferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayResponse"
                }
              }
            }
          },
          "401": {
            "description": "An authorization error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid amount, self-transfer, insufficient funds, or recipient not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/user/me/profile": {
      "get": {
        "tags": [
          "crate::routes::user"
        ],
        "operationId": "my_profile",
        "responses": {
          "200": {
            "description": "Aggregated profile for the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/user/me/sync-linked-role": {
      "post": {
        "tags": [
          "crate::routes::user"
        ],
        "operationId": "sync_linked_role",
        "responses": {
          "204": {
            "description": "Linked-role metadata pushed to Discord"
          },
          "401": {
            "description": "Auth or stored token missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Discord rejected the push or refresh",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/user/{id}/profile": {
      "get": {
        "tags": [
          "crate::routes::user"
        ],
        "operationId": "user_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated profile for the requested user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminUserDto": {
        "type": "object",
        "required": [
          "id",
          "boonbucks",
          "watched_time",
          "migrated",
          "role",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "boonbucks": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "last_message_sent": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "migrated": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "watched_time": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ConnectedYoutubeAccountDto": {
        "type": "object",
        "description": "A linked YouTube channel.",
        "required": [
          "channel_id",
          "channel_name"
        ],
        "properties": {
          "channel_id": {
            "type": "string"
          },
          "channel_name": {
            "type": "string"
          }
        },
        "examples": [
          {
            "channel_id": "UCxxxx",
            "channel_name": "Cool Channel"
          }
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "An error response containing details about the problem",
        "required": [
          "message",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A machine-readable error code identifying the error type.\n\nCommon values include \"Bad Request\", \"Unauthorized\", \"Internal Server Error\",\nand \"Not Found\".",
            "examples": [
              "Bad Request",
              "Unauthorized",
              "Internal Server Error"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable message describing the error.",
            "examples": [
              "Invalid authorization code"
            ]
          }
        },
        "examples": [
          {
            "error": "Unauthorized",
            "message": "Invalid authorization code"
          }
        ]
      },
...
      "PayRequest": {
        "type": "object",
        "description": "Pay boonbucks to another user",
        "required": [
          "recipient_id",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32",
            "description": "Amount of boonbucks to send. Must be > 0.",
            "examples": [
              50
            ]
          },
          "recipient_id": {
            "type": "integer",
            "format": "int64",
            "description": "Discord user ID of the recipient.",
            "examples": [
              675674657
            ]
          }
        },
        "examples": [
          {
            "amount": 50,
            "recipient_id": 675674657
          }
        ]
      },
      "PayResponse": {
        "type": "object",
        "description": "Result of a successful pay transfer",
        "required": [
          "amount",
          "sender_balance",
          "recipient_balance"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32",
            "description": "Amount transferred."
          },
          "recipient_balance": {
            "type": "integer",
            "format": "int32",
            "description": "Recipient's balance after the transfer."
          },
          "sender_balance": {
            "type": "integer",
            "format": "int32",
            "description": "Sender's balance after the transfer."
          }
        },
        "examples": [
          {
            "amount": 50,
            "recipient_balance": 1050,
            "sender_balance": 950
          }
        ]
      },
...
      "ProfileDto": {
        "type": "object",
        "description": "Aggregated profile view of a user.",
        "required": [
          "id",
          "boonbucks",
          "watched_time",
          "listening_hours",
          "position_by_hours",
          "position_by_boonbucks",
          "cans_added",
          "connected_youtube_accounts",
          "role",
          "permissions",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "boonbucks": {
            "type": "integer",
            "format": "int32"
          },
          "cans_added": {
            "type": "integer",
            "format": "int64",
            "description": "Number of cans this user has placed in Can Town."
          },
          "connected_youtube_accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectedYoutubeAccountDto"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "listening_hours": {
            "type": "integer",
            "format": "int32",
            "description": "`watched_time / 3600`, exposed for convenience."
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Effective permissions (role grants + per-user grants − per-user revokes)."
          },
          "position_by_boonbucks": {
            "type": "integer",
            "format": "int64",
            "description": "1-based rank by `boonbucks` descending."
          },
          "position_by_hours": {
            "type": "integer",
            "format": "int64",
            "description": "1-based rank by `watched_time` descending."
          },
          "role": {
            "type": "string",
            "description": "The user's primary role name."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "watched_time": {
            "type": "integer",
            "format": "int64",
            "description": "Watched (listening) time in seconds."
          }
        },
        "examples": [
          {
            "boonbucks": 1234,
            "cans_added": 25,
            "connected_youtube_accounts": [],
            "created_at": "2023-01-01T00:00:00Z",
            "id": 675674657,
            "listening_hours": 10,
            "permissions": [
              "use_minigames",
              "use_web_chat",
              "use_bot"
            ],
            "position_by_boonbucks": 7,
            "position_by_hours": 3,
            "role": "user",
            "updated_at": "2023-01-01T00:00:00Z",
            "username": "Caliborn",
            "watched_time": 36000
          }
        ]
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "boonbucks": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "migrated": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "watched_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "description": "A user of lumiRadio",
        "required": [
          "id",
          "watched_time",
          "boonbucks",
          "migrated",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "boonbucks": {
            "type": "integer",
            "format": "int32",
            "description": "The amount of boonbucks (virtual currency) the user has accumulated.\n\nBoonbucks can be earned through various activities on the platform and\ncan be used for purchases or interactions.",
            "examples": [
              100
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the user account was created.",
            "examples": [
              "2023-01-01T00:00:00Z"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The unique identifier of the user.",
            "examples": [
              675674657
            ]
          },
          "last_message_sent": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "The timestamp of the last message sent by the user.\n\nThis is `None` if the user has never sent a message.",
            "examples": [
              "2023-01-01T00:00:00Z"
            ]
          },
          "migrated": {
            "type": "boolean",
            "description": "Whether the user has migrated their data from the old radio bot.\n\nThis is `true` if the user has migrated their data from the old radio bot.",
            "examples": [
              true
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the user account was last updated.",
            "examples": [
              "2023-01-01T00:00:00Z"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "description": "The username of the user, if they have set one.\n\nThis is `None` if the user has not set a username yet.",
            "examples": [
              "Caliborn"
            ]
          },
          "watched_time": {
            "type": "integer",
            "format": "int64",
            "description": "The total watched time of the user, in seconds.\n\nThis represents the cumulative time the user has spent watching content\non the lumiRadio platform.",
            "examples": [
              1234567890
            ]
          }
        },
        "examples": [
          {
            "boonbucks": 100,
            "created_at": "2023-01-01T00:00:00Z",
            "id": 675674657,
            "last_message_sent": "2023-01-01T00:00:00Z",
            "migrated": true,
            "updated_at": "2023-01-01T00:00:00Z",
            "username": "Caliborn",
            "watched_time": 1234567890
          }
        ]
      },
...
    },
    "securitySchemes": {
      "user_api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "user_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}